#14 #15 Serialize EDTFField values into DB
Improve performance of EDTFField when populating model
instances from DB values by storing EDTF data in pickled
format, not as string values that need to be re-parsed
every time they are loaded.

This implementation is naive and could be improved with
a more sophisticated serialization approach for EDTF
fields, though it works for now.

This change also handles the issue in #15 where already-
parsed EDTF field values are re-parsed when an instance
is saved, unless there is a `natural_text_field` value
present to override and reset the EDTF field value.