#26 EDTFField can now store derived dates as floats
Update `EDTFField` to store derived upper/lower
static/float dates as float values when these
fields are defined as `DoubleField` in the target
model.

This change permits representation of arbitrary
dates not limited to 1 to 9999 AD like the
alternative `DateField` target fields.

Note that the stored value is not completely
accurate and, in my testing, may drift by a few
years for negative years in the thousands or
millions – which seems like a reasonable
limitation.