Fix parsing specified day, unspecified year & month
Parsing of partial unspecified dates with day
specified but year and month unspecified was
failing due to a grammar bug that caused the
'month' rule name to be nested.

Thus 'uuuu-uu-23' through `parse_edtf()` would
become 'uuuu-{'month': 'uu'}-23' which failed
subsequent processing.

This change fixes the over-nesting of the 'month'
data.