Fix handling of "level 2" intervals with uncertain dates
Check whether incoming lower/upper values are
single-item lists, and if so, use the first item.

This works around what I *think* is a bug in the
grammer that provides us with single-item lists of
`PartialUncertainOrApproximate` items for
lower/upper values. But I have not been able to
find and fix the fault in the grammar.