Implement pyproject.toml
Remove setup.cfg and setup.py in favor of pyproject.toml

Nose and Tox are next to go. Nose seems to be failing on Python 3.10, no need for it any more as pytest should cover its features. Likely no need for tox either as we are using the matrix strategy in Github Actions, though it could be useful for local dev