Add Django integration tests
This basic Django app shows how a user could create a model using the EDTFField and store data in it. The integration tests check that the EDTFField and associated fields (date_edtf_direct and date_display, in this case) work correctly.

There is a weird issue in test_date_display() where if we use an instance variable (self.event1, self.event2) the event.date_display property is available, but if we retrieve the object from the database it is not. I tried using TestEvent.objects.create() as well as the current method (make and then save an instance to no effect).

CI is set up to run the Django integration tests after Pytest. We could move to using pytest/django-pytest for these tests as well