Updated parser classes
I've had a pass at the Parser Classes file, but there are a lot of problems still to be sorted out.

I've added return types and argument types whereever it makes sense.

The "UncertainOrApproximate" class is a hot mess. There are boolean values with property and method calls associated with them, and I would be surprised if it actually works. However, it doesn't seem to be tested or implemented, so I can't figure out where to go from here.